+++ src/main/java/com/umc/mada/todo/repository/CategoryStatisticsVO.java
@@ -3,5 +3,5 @@ package com.umc.mada.todo.repository;
 public interface CategoryStatisticsVO {
     String getCategoryName();
     String getColor();
-    Double getRate();
+    Float getRate();
 }
